Quick hack to allow enough html parsing for testpilot to work fully
This allows either templating the translated strings, or parsing the
translated strings as html, but not both. In testpilot app.ftl, I was
able to convert all the strings over to the new nested Localized
approach without problems. However in testpilot experiments.ftl, we
programatically generate the string names and the html strings aren't
standard enough across experiments to remove the html from the strings
entirely, but we only need the translated strings to be parsed as html
and there are no template substitutions which also need to take place.
So, allowing either substitution of templates or parsing of html but
not both works as a hack, but might be too gross to allow as an
official api.